Extract RSpec shared groups to be used by side projects.
