+++ example/adapter/lib/new_adapter.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+class NewAdapter < DoSnapshot::Adapter::Abstract
+end
