+++ spec/do_snapshot/command_spec.rb
@@ -3,14 +3,14 @@
 require 'spec_helper'
 
 RSpec.describe DoSnapshot::Command do
-  include_context 'environment'
-  include_context 'uri_helpers'
+  include DoSnapshot::RSpec::Environment
+  include DoSnapshot::RSpec::UriHelpers
 
   subject(:cmd)     { DoSnapshot::Command.new }
   subject(:log)     { DoSnapshot::Log }
 
   describe 'V2' do
-    include_context 'api_v2_helpers'
+    include DoSnapshot::RSpec::ApiV2Helpers
 
     describe '.snap' do
       context 'when success' do
